refactor: move permission constants outside
