+++ src/core/utils/constants.ts
@@ -14,6 +14,23 @@ export enum WorkspacePermissionMode {
   LibraryWrite = 'library_write',
 }
 
+export enum PermissionModeId {
+  Read = 'read',
+  ReadAndWrite = 'read+write',
+  Admin = 'admin',
+}
+
+export const OptionIdToWorkspacePermissionModesMap: {
+  [key: string]: WorkspacePermissionMode[];
+} = {
+  [PermissionModeId.Read]: [WorkspacePermissionMode.LibraryRead, WorkspacePermissionMode.Read],
+  [PermissionModeId.ReadAndWrite]: [
+    WorkspacePermissionMode.LibraryWrite,
+    WorkspacePermissionMode.Read,
+  ],
+  [PermissionModeId.Admin]: [WorkspacePermissionMode.LibraryWrite, WorkspacePermissionMode.Write],
+};
+
 export const PUBLIC_WORKSPACE_ID = 'public';
 
 export const MANAGEMENT_WORKSPACE_ID = 'management';
